remove circular dependency by separating client into its own module
